:art: PIC-865: Corrected incorrect capitalisation of defendant names
