+++ mappings/case-list-full.json
@@ -53,7 +53,7 @@
             }
           ],
           "defendantSex": "F",
-          "defendantName": "Kara Ayers",
+          "defendantName": "Kara AYERS",
           "defendantAddress": {
             "line1": "91",
             "line2": "Barbey Street",
@@ -140,7 +140,7 @@
             }
           ],
           "defendantSex": "M",
-          "defendantName": "Guadalupe Hess",
+          "defendantName": "Guadalupe HESS",
           "defendantAddress": {
             "line1": "43",
             "line2": "Hunterfly Place",
@@ -231,7 +231,7 @@
             }
           ],
           "defendantSex": "F",
-          "defendantName": "Felicia Villarreal",
+          "defendantName": "FELICIA Villarreal",
           "defendantAddress": {
             "line1": "37",
             "line2": "Maple Avenue",
@@ -1720,7 +1720,7 @@
             }
           ],
           "defendantSex": "F",
-          "defendantName": "Amanda Mckay",
+          "defendantName": "Amanda McKay",
           "defendantAddress": {
             "line1": "42",
             "line2": "Ivan Court",
