+++ server/utils/nunjucksSetup.js
@@ -1,51 +1,46 @@
 const nunjucks = require('nunjucks')
-const moment = require('moment')
 
 module.exports = (app, path) => {
-  const njkEnv = nunjucks.configure([
+  const env = nunjucks.configure([
     path.join(__dirname, '../../node_modules/govuk-frontend/'),
     path.join(__dirname, '../../node_modules/@ministryofjustice/frontend/'),
-    path.join(__dirname, '../views')
+    path.join(__dirname, '../../server/views')
   ], {
     autoescape: true,
     express: app,
     watch: process.env.WATCH_TEMPLATES
   })
 
-  njkEnv.addFilter('findError', (array, formFieldId) => {
-    const item = array.find(error => error.href === `#${formFieldId}`)
-    if (item) {
-      return {
-        text: item.text
+  env.addFilter('json', function (value) {
+    return JSON.parse(value)
+  })
+
+  env.addFilter('limit', (arr, limit) => {
+    return arr.slice(0, limit)
+  })
+
+  env.addFilter('properCase', (name) => {
+    return ('' + name.replace(/[^\s\-']+[\s\-']*/g, (word) => {
+      return word.charAt(0).toUpperCase() + word.substr(1).toLowerCase()
+    }).replace(/\b(Van|De|Der|Da|Von)\b/g, (nobiliaryParticle) => {
+      return nobiliaryParticle.toLowerCase()
+    }).replace(/Mc(.)/g, (match, letter3) => {
+      return 'Mc' + letter3.toUpperCase()
+    }))
+  })
+
+  env.addFilter('markMatches', (matchString, sourceString) => {
+    const sourceSplit = sourceString.split(' ').map(item => {
+      return item.replace(',', '').toLowerCase()
+    })
+    const filteredArr = matchString.split(' ').map(item => {
+      let hasComma = false
+      if (item.indexOf(',') !== -1) {
+        item = item.replace(',', '')
+        hasComma = true
       }
-    }
-    return null
-  })
-
-  njkEnv.addFilter('findErrors', (errors, formFieldIds) => {
-    const fieldIds = formFieldIds.map(field => `#${field}`)
-    const errorIds = errors.map(error => error.href)
-    const firstPresentFieldError = fieldIds.find(fieldId => errorIds.includes(fieldId))
-    if (firstPresentFieldError) {
-      return { text: errors.find(error => error.href === firstPresentFieldError).text }
-    }
-    return null
-  })
-
-  njkEnv.addFilter('hasErrorWithPrefix', (errorsArray, prefixes) => {
-    const formattedPrefixes = prefixes.map(field => `#${field}`)
-    return errorsArray.some(error => formattedPrefixes.some(prefix => error.href.startsWith(prefix)))
-  })
-
-  njkEnv.addFilter('formatDate', (value, format) => {
-    return value ? moment(value).format(format) : null
-  })
-
-  njkEnv.addFilter('extractAttr', (array, key) => {
-    return array.map(item => item[key])
-  })
-
-  njkEnv.addFilter('isArray', value => {
-    return Array.isArray(value)
+      return (sourceSplit.includes(item.toLowerCase()) ? `<mark>${item}</mark>` : item) + (hasComma ? ',' : '')
+    })
+    return filteredArr.join(' ')
   })
 }
