+++ server/views/case-list.njk
@@ -33,7 +33,7 @@
         {{ offences.push('</ol>') }}
     {% endif %}
     {%- set tableRow = [
-            { text: ('<a href="/' + params.courtCode + '/case/' + item.caseNo + '/summary" class="pac-defendant-link govuk-link govuk-link--no-visited-state">' + item.defendantName | escape + '</a>') | safe },
+            { text: ('<a href="/' + params.courtCode + '/case/' + item.caseNo + '/summary" class="pac-defendant-link govuk-link govuk-link--no-visited-state">' + item.defendantName | escape | properCase + '</a>') | safe },
             { text: ('<div>' + ('<span class="moj-badge moj-badge--red pac-badge">Possible nDelius Record</span>' | safe if notMatched ) + ('<span class="moj-badge moj-badge--black pac-badge">Breach</span>' | safe if item.breach) + ('<span class="moj-badge moj-badge--black pac-badge">Sso</span>' | safe if item.suspendedSentenceOrder) + '</div>'+ (item.probationStatus | string | capitalize if not notMatched) + ('<span data-cy="previously-known-termination-date" class="govuk-caption-m">Order ended ' + moment(item.previouslyKnownTerminationDate, 'YYYY-MM-DD').format("D MMMM YYYY") | escape + '</span>' if item.previouslyKnownTerminationDate)) | safe },
             { text: offences | join('') | safe },
             { text: item.listNo },
