+++ server/views/match-manual.njk
@@ -47,7 +47,7 @@
             <dl class="govuk-summary-list">
                 <div class="govuk-summary-list__row">
                     <dt class="govuk-summary-list__key">Name</dt>
-                    <dd class="govuk-summary-list__value">{{ data.defendantName }}</dd>
+                    <dd class="govuk-summary-list__value">{{ data.defendantName | properCase }}</dd>
                 </div>
                 <div class="govuk-summary-list__row">
                     <dt class="govuk-summary-list__key">Date of birth</dt>
